modules/kafkaPersonConsumer/requirements.txt
GeoAlchemy2==0.8.4
kafka-python==2.0.2
marshmallow==3.7.1
psycopg2-binary==2.8.5
shapely==1.7.0
SQLAlchemy==1.3.19